Renamed test files as per review comments
