+++ data-prepper-expression/src/test/java/org/opensearch/dataprepper/expression/GenericExpressionEvaluator_ArithmeticIT.java
@@ -31,9 +31,10 @@ import static org.hamcrest.CoreMatchers.isA;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.sameInstance;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.CoreMatchers.not;
 import org.apache.commons.lang3.RandomStringUtils;
 
-class StringExpressionEvaluatorIT {
+class GenericExpressionEvaluator_ArithmeticIT {
 
     private AnnotationConfigApplicationContext applicationContext;
 
@@ -45,13 +46,13 @@ class StringExpressionEvaluatorIT {
     }
 
     @Test
-    void testStringExpressionEvaluatorBeanAvailable() {
+    void testArithmeticExpressionEvaluatorBeanAvailable() {
         final GenericExpressionEvaluator evaluator = applicationContext.getBean(GenericExpressionEvaluator.class);
         assertThat(evaluator, isA(GenericExpressionEvaluator.class));
     }
 
     @Test
-    void testStringExpressionEvaluatorBeanSingleton() {
+    void testArithmeticExpressionEvaluatorBeanSingleton() {
         final GenericExpressionEvaluator instanceA = applicationContext.getBean(GenericExpressionEvaluator.class);
         final GenericExpressionEvaluator instanceB = applicationContext.getBean(GenericExpressionEvaluator.class);
         assertThat(instanceA, sameInstance(instanceB));
@@ -72,10 +73,10 @@ class StringExpressionEvaluatorIT {
 
     @ParameterizedTest
     @MethodSource("validExpressionArguments")
-    void testStringExpressionEvaluator(final String expression, final Event event, final String expected, final Class expectedClass) {
+    void testArithmeticExpressionEvaluator(final String expression, final Event event, final Number expected, final Class expectedClass) {
         final GenericExpressionEvaluator evaluator = applicationContext.getBean(GenericExpressionEvaluator.class);
 
-        final String actual = (String)evaluator.evaluate(expression, event);
+        final Number actual = (Number)evaluator.evaluate(expression, event);
 
         assertThat(actual, is(expected));
         assertThat(actual, instanceOf(expectedClass));
@@ -83,23 +84,23 @@ class StringExpressionEvaluatorIT {
 
     @ParameterizedTest
     @MethodSource("validExpressionArguments")
-    void testStringExpressionEvaluatorWithMultipleThreads(final String expression, final Event event, final String expected, final Class expectedClass) {
+    void testArithmeticExpressionEvaluatorWithMultipleThreads(final String expression, final Event event, final Number expected, final Class expectedClass) {
         final GenericExpressionEvaluator evaluator = applicationContext.getBean(GenericExpressionEvaluator.class);
 
         final int numberOfThreads = 50;
         final ExecutorService executorService = Executors.newFixedThreadPool(numberOfThreads);
 
-        List<String> evaluationResults = Collections.synchronizedList(new ArrayList<>());
+        List<Number> evaluationResults = Collections.synchronizedList(new ArrayList<>());
 
         for (int i = 0; i < numberOfThreads; i++) {
-            executorService.execute(() -> evaluationResults.add((String)evaluator.evaluate(expression, event)));
+            executorService.execute(() -> evaluationResults.add((Number)evaluator.evaluate(expression, event)));
         }
 
         await().atMost(5, TimeUnit.SECONDS)
                 .until(() -> evaluationResults.size() == numberOfThreads);
 
         assertThat(evaluationResults.size(), equalTo(numberOfThreads));
-        for (String evaluationResult : evaluationResults) {
+        for (Number evaluationResult : evaluationResults) {
             assertThat(evaluationResult, equalTo(expected));
             assertThat(evaluationResult, instanceOf(expectedClass));
         }
@@ -107,20 +108,23 @@ class StringExpressionEvaluatorIT {
 
     @ParameterizedTest
     @MethodSource("invalidExpressionArguments")
-    void testStringExpressionEvaluatorInvalidInput(final String expression, final Event event, final Class expectedClass) {
+    void testArithmeticExpressionEvaluatorInvalidInput(final String expression, final Event event, final Class expectedClass) {
         final GenericExpressionEvaluator evaluator = applicationContext.getBean(GenericExpressionEvaluator.class);
-
-        final Object result = evaluator.evaluate(expression, event);
+        Object result = evaluator.evaluate(expression, event);
         assertThat(result, not(instanceOf(expectedClass)));
     }
 
     private static Stream<Arguments> validExpressionArguments() {
         Random random = new Random();
+        int randomInt = random.nextInt(10000);
+        float randomFloat = random.nextFloat();
         int testStringLength = random.nextInt(30);
         String testString = RandomStringUtils.randomAlphabetic(testStringLength);
         return Stream.of(
-                Arguments.of("\""+testString+"\"", event("{}"), testString, String.class),
-                Arguments.of("/status_message", event("{\"status_message\": \""+testString+"\"}"), testString, String.class)
+                Arguments.of(""+randomInt, event("{}"), randomInt, Integer.class),
+                Arguments.of("/status_code", event("{\"status_code\": "+randomInt+"}"), randomInt, Integer.class),
+                Arguments.of("/status_code", event("{\"status_code\": "+randomFloat+"}"), randomFloat, Float.class),
+                Arguments.of("length(/message)", event("{\"message\": \""+testString+"\"}"), testString.length(), Integer.class)
         );
     }
 
@@ -128,11 +132,12 @@ class StringExpressionEvaluatorIT {
         Random random = new Random();
         int testStringLength = random.nextInt(10);
         String testString = RandomStringUtils.randomAlphabetic(testStringLength);
-        int randomInt = random.nextInt(10000);
         return Stream.of(
-                Arguments.of("/missing", event("{}"), String.class),
-                Arguments.of("/value", event("{\"value\": "+randomInt+"}"), String.class),
-                Arguments.of("length(/message)", event("{\"message\": \""+testString+"\"}"), String.class)
+                Arguments.of("/missing", event("{}"), Integer.class),
+                Arguments.of("/message", event("{\"message\": \""+testString+"\"}"), Integer.class),
+                Arguments.of("/status", event("{\"status\": true}"), Integer.class),
+                Arguments.of("/status", event("{\"status\": 5.55}"), Integer.class),
+                Arguments.of("/status", event("{\"status\": 200}"), Float.class)
         );
     }
 
@@ -140,4 +145,3 @@ class StringExpressionEvaluatorIT {
         return JacksonEvent.builder().withEventType("event").withData(data).build();
     }
 }
-
